requirements.txt
streamlit
google-generativeai
firecrawl-py